+++ src/components/Buttons/Buttons.jsx
@@ -2,8 +2,20 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button } from '@material-ui/core';
 
-export const SortButton = ({ alphabeticalSort, lengthSort }) => (
+
+export const Buttons = ({ reverse, alphabeticalSort, lengthSort, reset }) => (
   <>
+    <Button
+      className="buttons__button button-reverse"
+      variant="contained"
+      color="primary"
+      onClick={() => {
+        reverse();
+      }}
+    >
+      reverse
+    </Button>
+
     <Button
       className="buttons__button button-sort"
       variant="contained"
@@ -25,10 +37,23 @@ export const SortButton = ({ alphabeticalSort, lengthSort }) => (
     >
       sort by length
     </Button>
+
+    <Button
+      className="buttons__button button-reset"
+      variant="contained"
+      color="primary"
+      onClick={() => {
+        reset();
+      }}
+    >
+      reset
+    </Button>
   </>
 );
 
-SortButton.propTypes = {
+Buttons.propTypes = {
+  reverse: PropTypes.func.isRequired,
   alphabeticalSort: PropTypes.func.isRequired,
   lengthSort: PropTypes.func.isRequired,
-};
+  reset: PropTypes.func.isRequired,
+}
