+++ src/components/Buttons/index.js
@@ -1,0 +1,1 @@
+export * from './Buttons';
