--- src/components/ResetButton/ResetButton.jsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Button } from '@material-ui/core';
-
-export const ResetButton = ({ reset }) => (
-  <Button
-    className="buttons__button button-reset"
-    variant="contained"
-    color="primary"
-    onClick={() => {
-      reset();
-    }}
-  >
-    reset
-  </Button>
-);
-
-ResetButton.propTypes = {
-  reset: PropTypes.func.isRequired,
-};
