--- src/components/SortButton/index.js
@@ -1,1 +1,0 @@
-export * from './SortButton';
